# Decouple `size` dictionary from `EinExprs`