--- a/src/Counters.jl
+++ b/src/Counters.jl
@@ -3,11 +3,11 @@
 
 Count the number of mathematical operations will be performed by the contraction of the root of the `path` tree.
 """
-flops(expr::EinExpr) =
+flops(expr::EinExpr, sizedict) =
     if length(expr.args) == 0 || length(expr.args) == 1 && isempty(suminds(expr))
         0
     else
-        mapreduce(Base.Fix1(size, expr), *, Iterators.flatten((head(expr), suminds(expr))), init = one(BigInt))
+        mapreduce(i -> sizedict[i], *, Iterators.flatten((head(expr), suminds(expr))), init = one(BigInt))
     end
 
 """
@@ -15,11 +15,13 @@ flops(expr::EinExpr) =
 
 Count the amount of memory that will be freed after performing the contraction of the root of the `path` tree.
 """
-removedsize(expr::EinExpr) = mapreduce(prod ∘ size, +, expr.args) - prod(size(expr))
+function removedsize(expr::EinExpr, sizedict)
+    mapreduce(prod ∘ Base.Fix2(size, sizedict), +, expr.args) - prod(size(expr, sizedict))
+end
 
 """
     removedrank(path::EinExpr)
 
 Count the rank reduction after performing the contraction of the root of the `path` tree.
 """
-removedrank(expr::EinExpr) = mapreduce(ndims, max, expr.args) - ndims(expr)
+removedrank(expr::EinExpr, _) = mapreduce(ndims, max, expr.args) - ndims(expr)