--- a/src/Optimizers/Exhaustive.jl
+++ b/src/Optimizers/Exhaustive.jl
@@ -22,23 +22,31 @@ The algorithm has a ``\mathcal{O}(n!)`` time complexity if `outer = true` and ``
     outer::Bool = false
 end
 
-function einexpr(config::Exhaustive, path; cost = BigInt(0))
-    leader = Ref{NamedTuple{(:path, :cost),Tuple{EinExpr,BigInt}}}((;
-        path = einexpr(Naive(), path),
-        cost = mapreduce(config.metric, +, Branches(einexpr(Naive(), path), inverse = true), init = BigInt(0))::BigInt,
-    ))
-    cache = Dict{Vector{Symbol},BigInt}()
-    __einexpr_exhaustive_it(path, cost, config.metric, config.outer, leader, cache)
-    return leader[].path
-end
+function einexpr(config::Exhaustive, path, sizedict; cost = BigInt(0))
+    metric = Base.Fix2(config.metric, sizedict)
 
-function __einexpr_exhaustive_it(path, cost, metric, outer, leader, cache)
-    if length(path.args) == 1
-        # remove identity einsum (i.e. "i...->i...")
-        path = path.args[1]
+    leader = (; path = einexpr(Naive(), path), cost = mapreduce(config.metric, +, Branches(einexpr(Naive(), path))))
+    cache = Dict{Vector{ImmutableVector{Symbol,Vector{Symbol}}},BigInt}()
 
-        leader[] = (; path, cost = mapreduce(metric, +, Branches(path, inverse = true), init = BigInt(0))::BigInt)
-        return
+    function __einexpr_iterate(path, cost)
+        if length(path.args) <= 2
+            leader = (; path = path, cost = mapreduce(metric, +, Branches(path)))
+            return
+        end
+
+        for (i, j) in combinations(args(path), 2)
+            !config.outer && isdisjoint(head(i), head(j)) && continue
+            candidate = sum([i, j], skip = path.head ∪ hyperinds(path))
+
+            # prune paths based on metric
+            new_cost = cost + get!(cache, head.(candidate.args)) do
+                metric(candidate)
+            end
+            new_cost >= leader.cost && continue
+
+            new_path = EinExpr(head(path), [candidate, filter(∉([i, j]), args(path))...])
+            __einexpr_iterate(new_path, new_cost)
+        end
     end
 
     for (i, j) in combinations(args(path), 2)