--- a/src/Optimizers/Exhaustive.jl
+++ b/src/Optimizers/Exhaustive.jl
@@ -49,7 +49,6 @@ function __einexpr_exhaustive_it(
 
     for (i, j) in combinations(args(path), 2)
         !outer && isdisjoint(head(i), head(j)) && continue
-        # candidate = sum([i, j]; skip = hashyperinds ? path.head ∪ hyperinds(path) : path.head)
         candidate = sum(i, j; skip = hashyperinds ? path.head ∪ hyperinds(path) : path.head)
 
         # prune paths based on metric