--- a/src/Optimizers/Exhaustive.jl
+++ b/src/Optimizers/Exhaustive.jl
@@ -47,7 +47,7 @@ function __einexpr_exhaustive_it(
         return
     end
 
-    for (i, j) in combinations(args(path), 2)
+    for (i, j) in combinations(path.args, 2)
         !outer && isdisjoint(head(i), head(j)) && continue
         candidate = sum(i, j; skip = hashyperinds ? path.head ∪ hyperinds(path) : path.head)
 
@@ -57,7 +57,7 @@ function __einexpr_exhaustive_it(
         end
         new_cost >= leader[].cost && continue
 
-        new_path = SizedEinExpr(EinExpr(head(path), [candidate, filter(∉([i, j]), args(path))...]), path.size) # sum([candidate, filter(∉([i, j]), args(path))...], skip = path.head)
+        new_path = SizedEinExpr(EinExpr(head(path), [candidate, filter(∉([i, j]), path.args)...]), path.size) # sum([candidate, filter(∉([i, j]), args(path))...], skip = path.head)
         __einexpr_exhaustive_it(new_path, new_cost, metric, outer, leader; cache, hashyperinds)
     end
 end