--- a/src/Optimizers/KaHyPar.jl
+++ b/src/Optimizers/KaHyPar.jl
@@ -6,7 +6,7 @@ using Suppressor
 @kwdef struct HyPar <: Optimizer
     parts::Int = 2
     imbalance::Float32 = 0.03
-    stop::Function = <=(2) ∘ length ∘ Base.Fix2(getfield, :args)
+    stop::Function = <=(2) ∘ length ∘ Base.Fix2(getproperty, :args)
     configuration::Union{Nothing,Symbol,String} = nothing
     edge_scaler::Function = Base.Fix1(*, 1000) ∘ Int ∘ round ∘ log2
     vertex_scaler::Function = Base.Fix1(*, 1000) ∘ Int ∘ round ∘ log2
@@ -25,7 +25,7 @@ function EinExprs.einexpr(config::HyPar, path)
 
     # NOTE indices in `inds` should be in the same order as unique indices appear by iterating on `path.args` because `∪` retains order
     edge_weights = map(config.edge_scaler ∘ Base.Fix1(size, path), inds)
-    vertex_weights = map(config.vertex_scaler ∘ length, path.args)
+    vertex_weights = map(config.vertex_scaler ∘ length, args(path))
 
     hypergraph = KaHyPar.HyperGraph(incidence_matrix, vertex_weights, edge_weights)
 
@@ -36,13 +36,13 @@ function EinExprs.einexpr(config::HyPar, path)
         configuration = config.configuration,
     )
 
-    args = map(unique(partitions)) do partition
+    _args = map(unique(partitions)) do partition
         selection = partitions .== partition
-        count(selection) == 1 && return only(path.args[selection])
+        count(selection) == 1 && return only(args(path)[selection])
 
-        expr = sum(path.args[selection], skip = path.head)
+        expr = sum(args(path)[selection], skip = path.head)
         einexpr(config, expr)
     end
 
-    return EinExpr(path.head, args)
+    return sum(_args, skip = path.head)
 end