When manipulating `EinExpr`s (e.g. creating, summing, ...), creation and merge of `size` dictionaries incurs into a large amount of allocations.

This PR decouples `size` dictionary from `EinExpr` and introduces `SizedEinExpr`: a `EinExpr` wrapper with a global index size dictionary.

```julia

julia> using EinExprs, BenchmarkTools

expr = sum([
    EinExpr([:j, :b, :i, :h], Dict(i => 2 for i in [:j, :b, :i, :h])),
    EinExpr([:a, :c, :e, :f], Dict(i => 2 for i in [:a, :c, :e, :f])),
    EinExpr([:j], Dict(i => 2 for i in [:j])),
    EinExpr([:e, :a, :g], Dict(i => 2 for i in [:e, :a, :g])),
    EinExpr([:f, :b], Dict(i => 2 for i in [:f, :b])),
    EinExpr([:i, :h, :d], Dict(i => 2 for i in [:i, :h, :d])),
    EinExpr([:d, :g, :c], Dict(i => 2 for i in [:d, :g, :c])),
])

@benchmark path = einexpr(Exhaustive(), expr)
```

### Before
```julia
BenchmarkTools.Trial: 2442 samples with 1 evaluation.
 Range (min … max):  1.618 ms … 7.713 ms  ┊ GC (min … max):  0.00% … 42.34%
 Time  (median):     1.674 ms             ┊ GC (median):     0.00%
 Time  (mean ± σ):   2.046 ms ± 1.072 ms  ┊ GC (mean ± σ):  13.91% ± 16.30%

  █▆▂                                               ▁▁▁▁▁    
  ████▇▆▄▃▄▁▁▁▁▁▃▃▁▃▁▁▁▁▁▁▃▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▇██████▇ █
  1.62 ms     Histogram: log(frequency) by time     5.37 ms <

 Memory estimate: 4.51 MiB, allocs estimate: 65772.
```

### After
```julia
BenchmarkTools.Trial: 9124 samples with 1 evaluation.
 Range (min … max):  400.167 μs …  14.535 ms  ┊ GC (min … max):  0.00% … 66.90%
 Time  (median):     425.625 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   546.911 μs ± 966.583 μs  ┊ GC (mean ± σ):  11.38% ±  6.06%

  █▂                                                            ▁
  ██▆▅▄▄▃▁▃▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▃▇▆ █
  400 μs        Histogram: log(frequency) by time       8.27 ms <

 Memory estimate: 621.73 KiB, allocs estimate: 11344.
```