@mofeing 2023-12-28
`Greedy` also exhibits a similar speedup and memory usage reduction.

### Before
```julia
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.696 μs … 309.470 μs  ┊ GC (min … max): 0.00% … 96.32%
 Time  (median):     4.935 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.269 μs ±   9.726 μs  ┊ GC (mean ± σ):  6.01% ±  3.20%

                  ▁▃▄ ▆▇█▇ ▅▃▃                                 
  ▁▁▁▁▁▁▁▁▂▁▃▃▄▅▃▇███▇█████████▄▆▅▃▃▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  4.7 μs          Histogram: frequency by time        5.33 μs <

 Memory estimate: 7.27 KiB, allocs estimate: 160.
```

### After
```julia
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.913 μs … 252.412 μs  ┊ GC (min … max): 0.00% … 97.77%
 Time  (median):     1.988 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.117 μs ±   5.519 μs  ┊ GC (mean ± σ):  5.74% ±  2.18%

               ▁▂▄ ▅▇▇▇ █▇▆▅ ▄▃▂▂                              
  ▂▂▂▂▂▁▃▃▄▅▁▆████▁████▁████▁█████▁▇▆▅▅▁▄▄▃▃▁▃▂▂▂▂▁▂▂▂▂▁▂▂▂▂▂ ▄
  1.91 μs         Histogram: frequency by time        2.11 μs <

 Memory estimate: 2.39 KiB, allocs estimate: 43.
```